Debug

+++ src/arguments.py
@@ -51,7 +51,7 @@ def get_args_parser():
                         help="Batch size per GPU/CPU for PET evaluation.")
     parser.add_argument('--gradient_accumulation_steps', type=int, default=1,
                         help="Number of updates steps to accumulate before performing a backward/update pass in PET.")
-    parser.add_argument("--num_train_epochs", default=3, type=float,
+    parser.add_argument("--num_train_epochs", default=3, type=int,
                         help="Total number of training epochs to perform in PET.")
     parser.add_argument("--max_steps", default=-1, type=int,
                         help="If > 0: set total number of training steps to perform in PET. Override num_train_epochs.")
